Fix syslog fputs_unlocked namespace (bug 18530).
syslog (XSI POSIX) brings in references to fputs_unlocked (not
POSIX).  This patch fixes this by making fputs_unlocked into a weak
alias for __fputs_unlocked and using __fputs_unlocked as needed.  (No
linknamespace test XFAILs are removed because there are other failures
from syslog as well.)

Tested for x86_64 and x86 (testsuite, and comparison of disassembly of
installed stripped shared libraries).  Disassembly of installed
stripped shared libraries is unchanged on x86_64; on x86, I see some
small changes to instruction ordering and register choice, with no
apparent reason for such changes to be related to this patch, but they
also seem completely harmless with no change to code size.

	[BZ #18530]
	* libio/iofputs.c [!_IO_MTSAFE_IO] (__fputs_unlocked): Define as
	strong alias of _IO_fputs.  Use libc_hidden_def.
	* libio/iofputs_u.c (fputs_unlocked): Rename to __fputs_unlocked
	and define as weak alias of __fputs_unlocked.  Use
	libc_hidden_weak.
	* include/stdio.h (__fputs_unlocked): Declare.  Use
	libc_hidden_proto.
	* misc/syslog.c (__vsyslog_chk): Call __fputs_unlocked instead of
	fputs_unlocked.