Fix timezone tests run in parallel.
The Makefile rule is called one time for target %/UTC or %/Universal.
If the tests are run in parallel, UTC or Universal can trigger the rule.
This patch patch generates the second test-result file as a copy of the
generated one.
Thus the testsuite can rely on the existence of the test-result files
regardless of which target has triggered the rule.

ChangeLog:

	* timezone/Makefile (%/UTC %/Universal):
	Generate test-result files for UTC and Universal.